// client/src/types.ts
export type RegionRouting =
  | 'la1'
  | 'la2'
  | 'na1'
  | 'euw1'
  | 'eun1'
  | 'kr'
  | 'br1'
  | 'oc1'
  | 'jp1'
  | 'tr1'
  | 'ru'
  | 'ph2'
  | 'sg2'
  | 'th2'
  | 'tw2'
  | 'vn2';
// ...
export interface TournamentConfig {
  name: string;
  tagline: string;
  startDate: string;
  endDate: string;
  prizePool: string;
  rules: {
    title: string;
    description: string;
  }[];
  customItemsEnabled: boolean;
  defaultRegion: RegionRouting;
  activeApiKey?: string;
}